requirements.txt
fastapi
uvicorn
sqlalchemy
pgvector
psycopg
openai
mcp
agno
ollama
arxiv
ddgs
duckdb
exa-py
pypdf
matplotlib
reportlab
requests
wikipedia
python-dotenv
linear-python
linkup-sdk
tavily-sdk
serper-sdk
fastmcp